+++ cherrypic-api/src/main/java/org/cherrypic/domain/auth/service/AuthServiceImpl.java
@@ -2,7 +2,6 @@ package org.cherrypic.domain.auth.service;
 
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.cherrypic.auth.repository.RefreshTokenRepository;
 import org.cherrypic.domain.auth.dto.AccessTokenDto;
 import org.cherrypic.domain.auth.dto.RefreshTokenDto;
 import org.cherrypic.domain.auth.dto.request.IdTokenRequest;
@@ -11,13 +10,14 @@ import org.cherrypic.domain.auth.dto.response.TokenReissueResponse;
 import org.cherrypic.domain.auth.enums.OauthProvider;
 import org.cherrypic.domain.auth.exception.AuthErrorCode;
 import org.cherrypic.domain.auth.exception.AuthException;
+import org.cherrypic.domain.auth.repository.RefreshTokenRepository;
 import org.cherrypic.domain.auth.util.NicknameGenerator;
 import org.cherrypic.domain.member.exception.MemberErrorCode;
 import org.cherrypic.domain.member.exception.MemberException;
+import org.cherrypic.domain.member.repository.MemberRepository;
 import org.cherrypic.global.util.MemberUtil;
 import org.cherrypic.member.entity.Member;
 import org.cherrypic.member.entity.OauthInfo;
-import org.cherrypic.member.repository.MemberRepository;
 import org.springframework.security.oauth2.core.oidc.user.OidcUser;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
