+++ cherrypic-api/src/main/java/org/cherrypic/domain/auth/service/JwtTokenService.java
@@ -3,11 +3,11 @@ package org.cherrypic.domain.auth.service;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.cherrypic.auth.entity.RefreshToken;
-import org.cherrypic.auth.repository.RefreshTokenRepository;
 import org.cherrypic.domain.auth.dto.AccessTokenDto;
 import org.cherrypic.domain.auth.dto.RefreshTokenDto;
 import org.cherrypic.domain.auth.exception.AuthErrorCode;
 import org.cherrypic.domain.auth.exception.AuthException;
+import org.cherrypic.domain.auth.repository.RefreshTokenRepository;
 import org.cherrypic.domain.auth.util.JwtUtil;
 import org.cherrypic.member.entity.Member;
 import org.cherrypic.member.enums.MemberRole;
