+++ cherrypic-api/src/main/java/org/cherrypic/domain/payment/repository/PaymentRepository.java
@@ -1,4 +1,4 @@
-package org.cherrypic.payment.repository;
+package org.cherrypic.domain.payment.repository;
 
 import java.util.Optional;
 import org.cherrypic.payment.entity.Payment;
