+++ cherrypic-api/src/main/java/org/cherrypic/global/util/MemberUtil.java
@@ -5,8 +5,8 @@ import org.cherrypic.domain.auth.exception.AuthErrorCode;
 import org.cherrypic.domain.auth.exception.AuthException;
 import org.cherrypic.domain.member.exception.MemberErrorCode;
 import org.cherrypic.domain.member.exception.MemberException;
+import org.cherrypic.domain.member.repository.MemberRepository;
 import org.cherrypic.member.entity.Member;
-import org.cherrypic.member.repository.MemberRepository;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
