+++ cherrypic-api/src/test/java/org/cherrypic/payment/service/PaymentServiceTest.java
@@ -17,17 +17,17 @@ import okhttp3.MediaType;
 import okhttp3.ResponseBody;
 import org.cherrypic.IntegrationTest;
 import org.cherrypic.album.enums.AlbumPlan;
+import org.cherrypic.domain.member.repository.MemberRepository;
 import org.cherrypic.domain.payment.dto.request.PaymentReadyRequest;
 import org.cherrypic.domain.payment.exception.PaymentErrorCode;
 import org.cherrypic.domain.payment.exception.PaymentException;
+import org.cherrypic.domain.payment.repository.PaymentRepository;
 import org.cherrypic.domain.payment.service.PaymentService;
 import org.cherrypic.global.util.MemberUtil;
 import org.cherrypic.member.entity.Member;
 import org.cherrypic.member.entity.OauthInfo;
-import org.cherrypic.member.repository.MemberRepository;
 import org.cherrypic.payment.entity.Payment;
 import org.cherrypic.payment.enums.PaymentStatus;
-import org.cherrypic.payment.repository.PaymentRepository;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
